[Actions][Linux] Add more plugins to be built, fix all compile warnings and errors (#946)
* Fix all the build issues to make it possible to build more stuff in Actions

* Update Linux build template with new packages and plugins

Added additional packages and plugin options for build.

* Update ThunderNanoServices workflow to use development branch

* Fix PKG_CONFIG_PATH export for architecture

* Fix some compile errors on the MinSizeRel version

* Use PRIu64 to print long long unsigned int to work on both 32 and 64 bit

* Use the macro to suppress the release warning instead

* Update ThunderNanoServices workflow to use master branch

* Update examples/IOConnectorTest/CMakeLists.txt


* Update examples/IOConnectorTest/Module.h


* Define EXTERNAL in Module.h

---------

+++ examples/DynamicLoading/Yin/YinImplementation.cpp
@@ -200,7 +200,7 @@ namespace Plugin {
                 Core::File dataFile(_service->DataPath() + _dataFile);
 
                 if (dataFile.Open(true) == true) {
-                    const uint64_t size = std::min(dataFile.Size(), 4096ULL);
+                    const uint64_t size = std::min(dataFile.Size(), static_cast<uint64_t>(4096));
                     uint8_t *data = static_cast<uint8_t*>(ALLOCA(size + 1));
                     ASSERT(data != nullptr);
                     dataFile.Read(data, size);
